Add way to get coefficients from Affine
Add `as_coeffs` method to Affine so we can get the coefficients out
(necessary for conversion to Direct2D types, etc).

Also tweak documentation to explain the convention for rotation
(partially addressing #4) and the augmented matrix representation of
Affine, particularly that it's backwards from PostScript.

This doesn't address all of #4, it doesn't change any behavior.