Add barebones Size and Point types
These implementations are intentionally minimal; I would like to
avoid adding public API that we might want to break until we have
a clearer idea of how these types are to be used.

Currently, I think there are two options: either we treat these as
convenience types that are exposed so that clients that want them
can use them, and we make minimal changes to public API, _or_ we
take a larger look at where Vec2 is used and attempt to use these
types more thoroughly where it makes sense.