Refactor BezPathSegs into Segments
To accomodate for arbitrary iterators, the iterator needed a bit of adjustment. Previously, `BezPath::segments_for_slice` peeked at the first elements of the slice before iteration to find out the start position. Because we can't do that for arbitrary iterators (at least without resorting to `.peekable()`), the `next()` method has an additional branch that fills in the start position in the first iteration. As this branch is executed only once, it should, however, be quite easy on branch prediction.