Fancier solver for inverse arclength
Use the ITP method for solving inverse arclength, which converges more
quickly than bisection but is just as robust.

This patch makes the ITP solver available in `common` as well, as it is
likely to be of general use. It's also nice that the solver is factored
out from the arclength-specific logic.

Also use the simple analytical solution for inverse arc length of a
line.