Add a custom return type for BezPath::min_dist
This is part of a bit of cleanup before doing a release, and includes a
few other doc and license fixes. It also makes min_dist_param pub(crate)
instead of pub; mindist was not exported to this isn't available as
public API in any case. If someone specifically needs to use this at
some point we can consider making it public then.