Improve documentation and robustness of BezPath
Clearly document the correctness invariants of BezPath and the fact that
it can represent multiple subpaths. Add debug asserts to catch construction
of invalid paths.

Note: the test_elements_to_segments_starts_on_quad formerly constructed
an invalid path and expected particular behavior. This is changed to
expect a panic on debug assert.

Responsive to #211