Add div_exact method to match cu2qu rounding behavior
This adds a (currently `pub(crate)`) `Vec2::div_exact` method that performs component wise division and updates the cubic to quad spline code to make use of it. Allows us to match rounding behavior and output of fonttools cu2qu.

I'm unsure if there is a need to expose this method publicly.

Fixes #272