Implement TranslateScale * Affine (both directions)
As part of testing this, I also added `float_cmp` to dev dependencies
and implemented it for Point/Affine. I propose that we start to
standardize on this method for checking approximate equality.

Finally, I've added some MathML to the docs to make it so people can
understand the code. In the TranslateScale * Affine (and rev) case,
one probably doesn't need more docs, but having them proves the concept.