Draft of regularization method
This is one possible technique for improving the robustness of parallel curve generation - a method to detect cusps and "regularize" by perturbing the control points.

The implementation handles cases where the control points are coincident with the endpoints, and also simple cusps, but hasn't been explicitly tuned to handle cases where the control points are nearly co-linear.

It is likely that additional tuning of the "dimension" parameter would be worthwhile. As a first approximation, it can just be the accuracy parameter.