Better numerical robustness for stroking
A number of fixes to improve stroke robustness.

Infinite recursion on curve fit is avoided by checking whether subdivision has hit the floating point limit. That said, the quality of curve fit is poor when the input is discontinuous (which does happen for degenerate cubics with cusps, where all points are co-linear).

Thus, in stroking, the degenerate cases are detected and drawn with special-case logic (do_linear). This uses existing drawing capabilities to draw the degenerate cubic with lines and round joins.

Finally, do_join mistakenly detected a half-turn as zero turn, as it was only considering the cross product (which vanishes).

Fixes #300