Deprecate `BezPath::flatten`, prefer `flatten`.
The free-standing function is more useful as it works on a wider
range of types. The `BezPath` method has the documentation though
and is just a wrapper around the free-standing function.

Most people will want to use the free-standing function, so we
can encourage that and make that the place where the docs live.