add triangle shape (#350)
Implements (at least starts to) a triangle shape, im not sure if this is
useful enough to merge though.

Also while looking at the source i noticed that that places like
[here](https://github.com/linebender/kurbo/blob/main/src/rect.rs#L43)
use `Rect` instead of `Self` unlike other parts of the codebase. And
some parts of the documentation uses "\[\`Type\`\]" and other parts just
use "Type". Wasnt sure to make an issue or not but should there be a
cleanup of the codebase?

also there is alot of conversion between `Vec2` and `Point` right now.
is there a better way to handle these situations where i want to use
Vec2 methods on a Point but will end up converting right back to a Point
afterwards?

triangles have alot of opportunities for more specific methods so its
worth noting that this can be heavily expanded upon which could come
later. Aswell as this, i have made some decisions like using centroids
as the origin of the triangle
(https://web.archive.org/web/20131104015950/http://www.jimloy.com/geometry/centers.htm)
which might not be the best? let me know.

ill continue working on this, specifically the TODOs (see inlined todos)

---------

Co-authored-by: Richard Dodd <richard.o.dodd@gmail.com>
Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>