Numerically approximate ellipse perimeter (#383)
This implements the (truncated) infinite series for the ellipse
perimeter as described by Kummer (1837) and rediscovered by Linderholm
and Segal (1995), and the iterative arithmetic-geometric mean method.

In the common case of ellipses that are only moderately eccentric and
for "normal" accuracy (say, 0.001), the infinite series converges
quickly. The series is known at compile-time, truncated to the 6th
power. For a given ellipse and accuracy, a quick check is performed at
runtime, to determine whether the truncated series' approximation is
within the desired accuracy. If so, the truncated series is evaluated.

If the check determines the approximation is not good enough, the
problem is handed to the iterative arithmetic-geometric mean method.
This method converges quadratically for all cases.