Reduce number of operations in `Triangle::circumscribed_circle`
(On top of PR #387, as I'm removing a method here that is still used
until that PR is merged. I'll keep this in draft until then.)

This is based on the simplified form for Cartesian coordinates in
https://en.wikipedia.org/w/index.php?title=Circumcircle&oldid=1247179617#Cartesian_coordinates_2

Also adds a test to check the circle radius' sign is equal to the sign
of the triangle's area.