Add Linebender lint set v2. (#402)
Actually satisfying a good chunk of these lints is deferred and they are
in an `allow` block in `lib.rs`.

Still, I did satisfy some of the lints.

The `rand` method `gen` has to be a raw `r#gen` until they publish a new
version. They have that method renamed in
[`master`](https://github.com/rust-random/rand).