Fix documentation of cross product; add convenience Vec2 methods (#409)
I managed to get this wrong when I was trying to nail down the sign
conventions, and it's been bothering me since I noticed it.

The PR adds a couple of simple convenience methods I find myself using
quite a bit as I implement stroke expansion.

Note that a lot of methods can become const when we bump the MSRV, but
that should happen separately.