Add `min` and `max` methods to `Size`. (#412)
This pull request adds a `min` and `max` method to the `Size` type.

# Rational

I found myself getting often the minimum of two size component-wise and
thought I'd be nicer to have a `min` and `max` method added to the type
directly.

Those two methods go in pair with the already existing `.clamp` method,
which does exactly that, but with a minimum and a maximum bound.