Move benchmarks into the kurbo crate, and port to criterion (#468)
As far as I can tell, benches at the top level of a virtual workspace
don't actually get built or run, so this PR moves them down into
`kurbo/`.

I've also taken the liberty to port them to criterion, so that they can
be run with stable rust. This requires bumping MSRV to 1.66 (for
`std::hint::black_box`), but I went all the way up to 1.82 so that there
won't be merge conflicts with #427

This PR should have no user-visible impact. I don' think there's any
rush to get it in a patch release.