Rework `Affine::svd` to only require two parallel square roots (#499)
On main, `Affine::svd` requires three square roots to calculate both
scales, or two square roots to calculate a single scale (through
dead-code elimination). Additionally, on main the first square root is
nested in the next two.

We can rework the maths to get an expression that requires just two
parallel square roots to get both scales. In case we need just one scale
(e.g., for getting the semi-major axis of a transformed unit circle),
this should still be slightly faster, as due to square roots being
parallel, the computations can be pipelined.

This is based on and expands
https://github.com/linebender/vello/blob/54e2a47abd0a9b1ad8b172bbaffed97d1c2248d6/vello_shaders/shader/flatten.wgsl#L360-L367.
I've added a derivation as comment (though, granted, if I didn't know
the form of the end goal, I probably wouldn't have found that
factorization of the inner term).