Change `Line::nearest` implementation to be branchless (#505)
This is a branchless calculation of a point-onto-line-segment
projection. Also marks the method `#[inline]` and adds some testing.

I started looking at this because of the math we introduced in
https://github.com/linebender/vello/pull/1214. This might allow us to
introduce some tighter bounds there without introducing additional
branching.

This should be useful here regardless of whether the above happens.

---------

Co-authored-by: jneem <joeneeman@gmail.com>