Make the linearity test scale-invariant (#503)
Fixes #477.

The issue here is in the test that we use to check whether the analytic
solution is numerically stable: there's a term that's unstable when the
three control points are colinear. There's a TODO in the code about
tuning the constant for that test. I haven't tried to tune the constant,
because the bigger issue in #477 is that the test isn't scale-invariant.
So the un-tuned constant was fine when the control points were about 1
unit apart, but broke when they were about 1000 units apart.