Bring poly-cool into the kurbo repo (#502)
As discussed in the last office hours, this PR brings the `poly-cool`
crate into the `kurbo` repo in preparation for making it a `kurbo`
dependency. This is almost a copy of the last poly-cool release version,
except that I deleted some of the `#[doc(hidden)]` experiments.

- I bumped the MSRV to 1.85 because that's poly-cool's MSRV and I think
it makes sense just to have one (plus, that's how the CI is set up).
- Do we have a process for managing changelogs in a multi-crate repo? I
looked in linebender/xilem for an example, but I couldn't find a
changelog...
- What's the crates.io process? I tried adding `github:org:linebender`
as an owner but crates.io didn't like it. Should I just add the kurbo
owners individually?

---------

Co-authored-by: Tom Churchman <thomas@churchman.nl>